requirements.txt
django-debug-toolbar
django==2.2.10
django-filter==2.1.0
django-webpack-loader==0.6.0
django-cors-headers==3.0.2
mysqlclient==1.4.2.post1
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
requests
django-bulk-update==2.2.0
django_mysql==3.3.0
drf_yasg==1.17.1
celery==4.3.0
redis==3.3.4
django-redis-cache==2.0.0
django-redis==4.10.0
